+++ frontend/src/app/main/data/workspace/libraries.cljs
@@ -620,10 +620,11 @@
      ptk/WatchEvent
      (watch [it state _]
        (log/info :msg "UPDATE-COMPONENT of shape" :id (str id) :undo-group undo-group)
-       (let [page-id     (get state :current-page-id)
-             local-file  (wsh/get-local-file state)
-             container   (cph/get-container local-file :page page-id)
-             shape       (ctn/get-shape container id)]
+       (let [page-id       (get state :current-page-id)
+             local-file    (wsh/get-local-file state)
+             container     (cph/get-container local-file :page page-id)
+             shape         (ctn/get-shape container id)
+             components-v2 (features/active-feature? state :components-v2)]
 
          (when (ctk/instance-head? shape)
            (let [libraries (wsh/get-libraries state)
@@ -632,7 +633,7 @@
                  (-> (pcb/empty-changes it)
                      (pcb/set-undo-group undo-group)
                      (pcb/with-container container)
-                     (dwlh/generate-sync-shape-inverse libraries container id))
+                     (dwlh/generate-sync-shape-inverse libraries container id components-v2))
 
                  file-id   (:component-file shape)
                  file      (wsh/get-file state file-id)
@@ -713,8 +714,6 @@
        (rx/of (dwu/start-undo-transaction undo-id))
        (rx/map #(update-component-sync (:id %) file-id (uuid/next)) (rx/from shapes))
        (rx/of (dwu/commit-undo-transaction undo-id)))))))
-
-(declare sync-file-2nd-stage)
 
 (def valid-asset-types
   #{:colors :components :typographies})
@@ -803,44 +802,7 @@
               (rx/concat (rx/timer 3000)
                          (rp/cmd! :update-file-library-sync-status
                                   {:file-id file-id
-                                   :library-id library-id})))
-            (when (and (seq (:redo-changes library-changes))
-                       sync-components?)
-              (rx/of (sync-file-2nd-stage file-id library-id asset-id undo-group))))))))))
-
-(defn- sync-file-2nd-stage
-  "If some components have been modified, we need to launch another synchronization
-  to update the instances of the changed components."
-  ;; TODO: this does not work if there are multiple nested components. Only the
-  ;;       first level will be updated.
-  ;;       To solve this properly, it would be better to launch another sync-file
-  ;;       recursively. But for this not to cause an infinite loop, we need to
-  ;;       implement updated-at at component level, to detect what components have
-  ;;       not changed, and then not to apply sync and terminate the loop.
-  [file-id library-id asset-id undo-group]
-  (dm/assert! (uuid? file-id))
-  (dm/assert! (uuid? library-id))
-  (dm/assert! (or (nil? asset-id)
-                  (uuid? asset-id)))
-  (ptk/reify ::sync-file-2nd-stage
-    ptk/WatchEvent
-    (watch [it state _]
-      (log/info :msg "SYNC-FILE (2nd stage)"
-                :file (dwlh/pretty-file file-id state)
-                :library (dwlh/pretty-file library-id state))
-      (let [file    (wsh/get-file state file-id)
-            changes (reduce
-                     pcb/concat-changes
-                     (-> (pcb/empty-changes it)
-                         (pcb/set-undo-group undo-group))
-                     [(dwlh/generate-sync-file it file-id :components asset-id library-id state)
-                      (dwlh/generate-sync-library it file-id :components asset-id library-id state)])]
-
-        (log/debug :msg "SYNC-FILE (2nd stage) finished" :js/rchanges (log-changes
-                                                                       (:redo-changes changes)
-                                                                       file))
-        (when (seq (:redo-changes changes))
-          (rx/of (dch/commit-changes (assoc changes :file-id file-id))))))))
+                                   :library-id library-id}))))))))))
 
 (def ignore-sync
   "Mark the file as ignore syncs. All library changes before this moment will not