+++ frontend/src/app/main/ui/viewer/shapes.cljs
@@ -38,11 +38,10 @@
 
 (defn- find-relative-to-base-frame
   [shape objects overlays-ids base-frame]
-  (if (or (empty? overlays-ids) (nil? shape) (cph/root? shape))
-    base-frame
-    (if (contains? overlays-ids (:id shape))
-      shape
-      (find-relative-to-base-frame (cph/get-parent objects (:id shape)) objects overlays-ids base-frame))))
+  (cond
+    (cph/frame-shape? shape) shape
+    (or (empty? overlays-ids) (nil? shape) (cph/root? shape)) base-frame
+    :else (find-relative-to-base-frame (cph/get-parent objects (:id shape)) objects overlays-ids base-frame)))
 
 (defn- activate-interaction
   [interaction shape base-frame frame-offset objects overlays]
@@ -272,7 +271,6 @@
               :transform (gsh/transform-str shape)}])))
 
 
-
 ;; TODO: use-memo use-fn
 
 (defn generic-wrapper-factory
@@ -468,13 +466,13 @@
                   (mf/with-memo [objects]
                     (svg-raw-container-factory objects all-objects))]
               (when (and shape (not (:hidden shape)))
-          (let [shape (-> shape
-                          #_(gsh/transform-shape)
-                                (gsh/translate-to-frame frame))
-
-                      opts #js {:shape shape
-                                :objects objects
-                                :all-objects all-objects}]
+          (let [shape (if frame
+                        (gsh/translate-to-frame shape frame)
+                        shape)
+                
+                opts #js {:shape shape
+                          :objects objects
+                          :all-objects all-objects}]
                   (case (:type shape)
                     :frame   [:> frame-container opts]
                     :text    [:> text-wrapper opts]