+++ frontend/src/app/thumbnail_renderer.cljs
@@ -97,11 +97,12 @@
 (defn- svg-update-image!
   "Updates an image in an SVG to a Data URI."
   [image]
-  (when-let [href (dom/get-attribute image "href")]
+  (if-let [href (dom/get-attribute image "href")]
     (->> (fetch-as-data-uri href)
          (rx/map (fn [url]
                    (dom/set-attribute! image "href" url)
-                   image)))))
+                   image)))
+    (rx/empty)))
 
 (defn- svg-resolve-images!
   "Resolves all images in an SVG to Data URIs."
@@ -191,6 +192,7 @@
                            (rx/mapcat #(wapi/create-image-bitmap % #js {:resizeWidth width
                                                                         :resizeQuality "medium"}))
                            (rx/tap #(wapi/revoke-uri uri)))))
+
          (rx/mapcat bitmap->blob))))
 
 (defn- on-message